[qca-ssdk]: return ok directly if when ess-switch status value is disabled
For VP platform doesn't have ethernet ports, and it use same profile with
RUMI, so VP platform will be crash when loading dts and do later opearations.
In order to avoid this issue, we can add status=disabled to VP dts file,
and let it return OK directly when status value is disabled.

Change-Id: I74bc502ec4b3629c5df6e034173525f478a3588e
Signed-off-by: mingxinh <mingxinh@codeaurora.org>