[qca-ssdk]: fix the igmp crash issue
this issue is caused by the multi device, where the api macro always
uses the default device id 0 instead of the correct device id.

move the macros into the directory app, these macros are only used by
the app code and defined to the fal APIs if HSL_STANDALONG is not used.

Change-Id: I6c0bdc4867ea6b51c4f3f8ceae63454502b001df
Signed-off-by: Luo Jie <luoj@codeaurora.org>