[qca-ssdk]: fix ptp jitter issue
when the slave clock tracks the master clock, and the frequency offset
reaches to 0(the clock is synchronized completely), the calculated
fraction nanoseconds is 0, which leads to the increasement value
error, it is necessary to take this case into consideration.

this patch also configure the RTC clock to the synchronization mode,
which enhances the accurancy on the adjustment of the clock.

Change-Id: Iae68150cea945dd872e3cb1b14f726494785ee76
Signed-off-by: Luo Jie <luoj@codeaurora.org>