[qca-ssdk]: Force ipv6 checksum zero in ptp embeded mode
In RX embeded mode, the timestamp is inserted into the reserved
fields of ptp header, the checksum needs to be recalculated.

Currently the HW does not update the checksum of packet with the
timestamp embeded, so configuring the udp6 checksum zeroed to make
the packet received with the socket option UDP_NO_CHECK6_RX enabled.

Change-Id: I6025234ce247b0efc5d462cdca2badf312568571
Signed-off-by: Luo Jie <luoj@codeaurora.org>